refactor: make uni formula abstract
