+++ packages-experimental/uni-formula-ui/package.json
@@ -64,6 +64,7 @@
         "@univerjs/docs-ui": "workspace:*",
         "@univerjs/icons": "^0.1.72",
         "@univerjs/sheets-formula": "workspace:*",
+        "@univerjs/rpc": "workspace:*",
         "@univerjs/ui": "workspace:*",
         "@univerjs/uni-formula": "workspace:*",
         "clsx": ">=2.0.0",
@@ -78,6 +79,7 @@
         "@univerjs/icons": "^0.1.72",
         "@univerjs/shared": "workspace:*",
         "@univerjs/sheets-formula": "workspace:*",
+        "@univerjs/rpc": "workspace:*",
         "@univerjs/ui": "workspace:*",
         "@univerjs/uni-formula": "workspace:*",
         "clsx": "^2.1.1",
