+++ packages-experimental/uni-formula-ui/src/index.ts
@@ -15,3 +15,4 @@
  */
 
 export { UniverDocUniFormulaUIPlugin } from './uni-formula-ui.plugin';
+export { UniFormulaService } from './services/uni-formula.service'
