+++ packages-experimental/uni-formula-ui/src/uni-formula-ui.plugin.ts
@@ -16,10 +16,12 @@
 
 import type { Dependency } from '@univerjs/core';
 import { DependentOn, Inject, Injector, Plugin, UniverInstanceType } from '@univerjs/core';
-import { UniverDocUniFormulaPlugin } from '@univerjs/uni-formula';
+import { IUniFormulaService, UniverDocUniFormulaPlugin } from '@univerjs/uni-formula';
+
 import { DOC_FORMULA_UI_PLUGIN_NAME } from './const';
 import { DocFormulaPopupService } from './services/formula-popup.service';
 import { DocUniFormulaController } from './controllers/formula-input.controller';
+import { UniFormulaService } from './services/uni-formula.service';
 
 @DependentOn(UniverDocUniFormulaPlugin)
 export class UniverDocUniFormulaUIPlugin extends Plugin {
@@ -37,6 +39,7 @@ export class UniverDocUniFormulaUIPlugin extends Plugin {
         ([
             [DocUniFormulaController],
             [DocFormulaPopupService],
+            [IUniFormulaService, { useClass: UniFormulaService }]
         ] as Dependency[]).forEach((d) => injector.add(d));
     }
 }
